Fix issue with mulh and negative numbers that caused DOOM to crash.

+++ mini-rv32ima/mini-rv32ima.h
@@ -293,8 +293,8 @@ MINIRV32_DECORATE int32_t MiniRV32IMAStep( struct MiniRV32IMAState * state, uint
 						switch( (ir>>12)&7 ) //0x02000000 = RV32M
 						{
 							case 0b000: rval = rs1 * rs2; break; // MUL
-							case 0b001: rval = ((int64_t)rs1 * (int64_t)rs2) >> 32; break; // MULH
-							case 0b010: rval = ((int64_t)rs1 * (uint64_t)rs2) >> 32; break; // MULHSU
+							case 0b001: rval = ((int64_t)((int32_t)rs1) * (int64_t)((int32_t)rs2)) >> 32; break; // MULH
+							case 0b010: rval = ((int64_t)((int32_t)rs1) * (uint64_t)rs2) >> 32; break; // MULHSU
 							case 0b011: rval = ((uint64_t)rs1 * (uint64_t)rs2) >> 32; break; // MULHU
 							case 0b100: if( rs2 == 0 ) rval = -1; else rval = (int32_t)rs1 / (int32_t)rs2; break; // DIV
 							case 0b101: if( rs2 == 0 ) rval = 0xffffffff; else rval = rs1 / rs2; break; // DIVU
